♻️ refactor: 디렉토리 common 구조 변경 및 상수 파일 분리
