+++ feed-crawler/src/common/ai/ai.type.ts
@@ -1,0 +1,17 @@
+export interface ClaudeResponse {
+  tags: Record<string, number>;
+  summary: string;
+}
+
+export type FeedAIQueueItem = {
+  id: number;
+  content: string;
+  deathCount: number;
+  tagList?: string[];
+  summary?: string;
+};
+
+export interface AiSummaryRetryMessage {
+  feedId: number;
+  deathCount: number;
+}
