+++ feed-crawler/src/common/database/database.constant.ts
@@ -1,0 +1,1 @@
+export const CONNECTION_LIMIT = 50;
