+++ feed-crawler/src/common/redis/redis-access.ts
@@ -3,7 +3,7 @@ import { injectable } from 'tsyringe';
 import Redis, { ChainableCommander } from 'ioredis';
 import Redis_Mock from 'ioredis-mock';
 
-import logger from '@common/logger';
+import logger from '@common/logger/logger';
 
 @injectable()
 export class RedisConnection {
