+++ feed-crawler/src/common/redis/redis.constant.ts
@@ -1,0 +1,7 @@
+export const redisConstant = {
+  FEED_RECENT_INDEX_KEY: 'feed:recent:index',
+  FEED_AI_QUEUE: `feed:ai:queue`,
+  FEED_AI_RETRY_QUEUE: `feed:ai-retry:queue`,
+  FEED_AI_RETRY_LOCK: `feed:ai-retry:lock`,
+  FULL_FEED_CRAWL_QUEUE: `feed:full-crawl:queue`,
+};
